Move `Gossip` mempool type into new crate
It is required by the `Request` type, which will be moved next.
