+++ zebra-node-services/src/lib.rs
@@ -1,1 +1,3 @@
 //! The interfaces of some Zebra node services.
+
+pub mod mempool;
