+++ zebra-node-services/src/mempool.rs
@@ -1,0 +1,34 @@
+//! The Zebra mempool.
+//!
+//! A service that manages known unmined Zcash transactions.
+
+use zebra_chain::transaction::{UnminedTx, UnminedTxId};
+
+/// A gossiped transaction, which can be the transaction itself or just its ID.
+#[derive(Debug, Eq, PartialEq)]
+pub enum Gossip {
+    Id(UnminedTxId),
+    Tx(UnminedTx),
+}
+
+impl Gossip {
+    /// Return the [`UnminedTxId`] of a gossiped transaction.
+    pub fn id(&self) -> UnminedTxId {
+        match self {
+            Gossip::Id(txid) => *txid,
+            Gossip::Tx(tx) => tx.id,
+        }
+    }
+}
+
+impl From<UnminedTxId> for Gossip {
+    fn from(txid: UnminedTxId) -> Self {
+        Gossip::Id(txid)
+    }
+}
+
+impl From<UnminedTx> for Gossip {
+    fn from(tx: UnminedTx) -> Self {
+        Gossip::Tx(tx)
+    }
+}
