+++ zebrad/src/components/mempool/crawler/tests/prop.rs
@@ -10,6 +10,7 @@ use tokio::time;
 
 use zebra_chain::{parameters::Network, transaction::UnminedTxId};
 use zebra_network as zn;
+use zebra_node_services::mempool::Gossip;
 use zebra_state::ChainTipSender;
 use zebra_test::mock_service::{MockService, PropTestAssertion};
 
@@ -17,7 +18,6 @@ use crate::components::{
     mempool::{
         self,
         crawler::{Crawler, SyncStatus, FANOUT, RATE_LIMIT_DELAY},
-        downloads::Gossip,
         error::MempoolError,
         Config,
     },
